Move PvP harness to core; add tokenizer counter + in-harness MCTS baseline
Model prep (a minimal image shipping only core/ + trainer/model_prep/) needs to
run the eval harness to compute a format-aligned baseline, so the shared harness
moves out of validator/ into core/pvp/ — the same shared-code split already used
for memory/tools.

Harness -> core/pvp/ (zero validator imports, verified):
- bot.py, agents.py, chat.py, scoring.py moved; game_eval.py holds the shared
  _AGENT_REGISTRY + timeout/forfeit evaluate_bots wrapper; constants.py holds the
  harness PVP_* constants. validator/core/constants.py re-exports them and
  validator/evaluation/pvp/{bot,agents,chat,scoring}.py are back-compat shims, so
  existing imports keep working.
- game_runner.py keeps the eval-only matchup orchestration, importing from core.

Tokenizer-backed memory budgets:
- core/pvp/tokenizer_counter.py: HFTokenCounter + load_token_counter (whitespace
  fallback for non-repo ids / load failures). game_runner builds a per-player
  counter so slot budgets are real model tokens, not whitespace words.

In-harness MCTS baseline:
- core/pvp/baseline.py: run_mcts_baseline plays the model (tool-calling LLMBot)
  vs a pyspiel MCTSBot, same format as eval, no external MCTS-API server.
- trainer/model_prep/env_stats.py now computes baseline stats via run_mcts_baseline
  instead of POSTing to the external /evaluate server.
- model-prep.dockerfile gains open-spiel + openai (harness runtime deps).

Tests: test_pvp_tokenizer.py, test_pvp_baseline.py; existing tests updated for the
new module paths. Full PvP suite (150) green.
