+++ core/pvp/baseline.py
@@ -1,0 +1,146 @@
+"""In-harness MCTS baseline.
+
+Play the model (the tool-calling LLMBot — identical format to eval) against a
+pyspiel MCTSBot and score the result. This gives a baseline game-playing number
+measured in the SAME interaction format as the PvP eval, with no external
+MCTS-API server: the opponent is built in-process from pyspiel.
+
+Used by model prep to gauge how well a base model plays before training.
+"""
+
+import logging
+import random
+
+import numpy as np
+import pyspiel
+from open_spiel.python.algorithms import mcts
+from pydantic import BaseModel
+
+from core.constants import ENVIRONMENT_CONFIGS
+from core.constants import EnvironmentName
+from core.models.pvp_models import ChatCompletionConfig
+from core.models.pvp_models import ChatFn
+from core.models.pvp_models import GameOutcome
+from core.models.pvp_models import GameScoringContext
+from core.models.pvp_models import MemoryArea
+from core.pvp.memory import SlotMemory
+from core.pvp.tokenizer_counter import load_token_counter
+from core.pvp import constants as cst
+from core.pvp.bot import LLMBot
+from core.pvp.game_eval import _AGENT_REGISTRY
+from core.pvp.game_eval import _evaluate_game_with_timeout
+from core.pvp.scoring import determine_outcome
+
+
+logger = logging.getLogger(__name__)
+
+_MCTS_UCT_C = 2.0
+_DEFAULT_MCTS_SIMULATIONS = 50
+
+
+class MctsBaselineResult(BaseModel):
+    """Outcome counts for a model's baseline games against MCTS."""
+
+    wins: int = 0
+    draws: int = 0
+    losses: int = 0
+    num_games: int = 0
+
+    @property
+    def mean_score(self) -> float:
+        """Win=1, draw=0.5, loss=0, averaged — a [0, 1] baseline score."""
+        if self.num_games == 0:
+            return 0.0
+        return (self.wins + 0.5 * self.draws) / self.num_games
+
+
+def _mcts_simulations_for(env_name: EnvironmentName) -> int:
+    extra = ENVIRONMENT_CONFIGS[env_name].eval_payload_extra or {}
+    return int(extra.get("mcts_max_simulations", _DEFAULT_MCTS_SIMULATIONS))
+
+
+def _make_mcts_bot(game: pyspiel.Game, simulations: int, seed: int) -> mcts.MCTSBot:
+    evaluator = mcts.RandomRolloutEvaluator(n_rollouts=1, random_state=np.random.RandomState(seed))
+    return mcts.MCTSBot(
+        game,
+        uct_c=_MCTS_UCT_C,
+        max_simulations=simulations,
+        evaluator=evaluator,
+        random_state=np.random.RandomState(seed),
+    )
+
+
+def run_mcts_baseline(
+    env_name: EnvironmentName,
+    chat_fn: ChatFn,
+    config: ChatCompletionConfig,
+    num_games: int,
+    mcts_simulations: int | None = None,
+    base_seed: int = 0,
+) -> MctsBaselineResult:
+    """Play num_games of env_name as the model (LLMBot) vs MCTS; return outcome counts.
+
+    The model alternates seats for fairness and carries one long-term memory across
+    the games (it builds a read on the MCTS opponent, exactly as in a real matchup).
+    A model-side forfeit (timeout, repeated illegal moves, context overflow) scores
+    as a loss, mirroring eval.
+    """
+    agent = _AGENT_REGISTRY[env_name]()
+    env_config = ENVIRONMENT_CONFIGS[env_name]
+    simulations = mcts_simulations if mcts_simulations is not None else _mcts_simulations_for(env_name)
+    counter = load_token_counter(config.inference_model)
+    long_term = SlotMemory(cst.PVP_LONGTERM_MEM_SLOTS, cst.PVP_LONGTERM_SLOT_TOKENS, counter)
+
+    seed_rng = random.Random(base_seed)
+    result = MctsBaselineResult()
+
+    for i in range(num_games):
+        seed = seed_rng.randint(1, cst.PVP_SEED_RANGE_MAX)
+        task_rng = random.Random(seed)
+        task_id = task_rng.randint(env_config.task_id_min, env_config.task_id_max)
+        config_id = task_id % cst.PVP_CONFIG_ID_DIVISOR
+        game = pyspiel.load_game(agent.game_name, agent.generate_params(config_id))
+        game_type = game.get_type()
+
+        model_seat = i % 2
+        mcts_seat = 1 - model_seat
+        working = SlotMemory(cst.PVP_WORKING_MEM_SLOTS, cst.PVP_WORKING_SLOT_TOKENS, counter)
+        model_bot = LLMBot(
+            game=game,
+            player_id=model_seat,
+            chat_fn=chat_fn,
+            config=config,
+            agent=agent,
+            rng_seed=seed + model_seat,
+            memories={MemoryArea.WORKING: working, MemoryArea.LONG_TERM: long_term},
+        )
+        bots: list = [None, None]
+        bots[model_seat] = model_bot
+        bots[mcts_seat] = _make_mcts_bot(game, simulations, seed + mcts_seat)
+
+        state = game.new_initial_state()
+        evaluation = _evaluate_game_with_timeout(state, bots, seed)
+        outcome = determine_outcome(
+            GameScoringContext(
+                returns=evaluation.returns,
+                player_id=model_seat,
+                is_zero_sum=game_type.utility == pyspiel.GameType.Utility.ZERO_SUM,
+                min_utility=game.min_utility(),
+                max_utility=game.max_utility(),
+            )
+        )
+        if outcome == GameOutcome.WIN:
+            result.wins += 1
+        elif outcome == GameOutcome.LOSS:
+            result.losses += 1
+        else:
+            result.draws += 1
+        result.num_games += 1
+
+        model_bot.reflect(state, outcome)
+
+    logger.info(
+        "%s MCTS baseline: %d games, %d-%d-%d (W-D-L), score=%.3f",
+        env_name.value, result.num_games, result.wins, result.draws, result.losses, result.mean_score,
+    )
+    return result
