+++ dockerfiles/model-prep.dockerfile
@@ -3,7 +3,8 @@ FROM winglian/axolotl:main-20251113
 WORKDIR /app
 
 RUN TORCH_VER=$(python -c "import torch; print(torch.__version__)") && \
-    pip install --no-cache-dir "sglang[srt]" "torch==${TORCH_VER}" datasketch aiohttp python-dotenv textstat
+    pip install --no-cache-dir "sglang[srt]" "torch==${TORCH_VER}" datasketch aiohttp python-dotenv textstat \
+    "open-spiel==1.6.13" openai
 
 COPY trainer/model_prep/ trainer/model_prep/
 COPY core/ core/
