+++ trainer/model_prep/env_stats.py
@@ -1,12 +1,13 @@
 """
-Environment task stats: deploy model via SGLang, play episodes against env servers.
-Self-contained — no validator imports. SGLang helpers inlined from eval_environment.py.
+Environment task stats: deploy model via SGLang, play baseline games against MCTS.
+No validator imports (model-prep ships core/ only) — the in-harness MCTS baseline
+lives in core.pvp. SGLang helpers inlined from eval_environment.py.
 """
 
 import asyncio
+import functools
 import logging
 import os
-import random
 import signal
 import socket
 import statistics
@@ -18,6 +19,10 @@ import aiohttp
 from core.constants import EnvironmentName
 from core.models.model_prep_models import EnvBaselineStats
 from core.models.model_prep_models import EnvStats
+from core.models.pvp_models import ChatCompletionConfig
+from core.pvp.baseline import run_mcts_baseline
+from core.pvp.chat import chat_completion
+from core.pvp.chat import create_client
 from trainer.model_prep.stats import compute_weight_stats
 
 
@@ -108,81 +113,42 @@ def _build_env_stats(scores: list[float]) -> EnvStats:
     return EnvStats(num_episodes=0)
 
 
-def _sample_task_id(seed: int, task_id_min: int, task_id_max: int) -> int:
-    return random.Random(seed).randint(task_id_min, task_id_max)
-
-
-async def _play_episodes(
-    session: aiohttp.ClientSession,
+def _mcts_baseline_stats(
     env_name: EnvironmentName,
-    env_server_url: str,
     sglang_base_url: str,
     model_name: str,
     num_episodes: int,
-    task_id_min: int,
-    task_id_max: int,
     eval_payload_extra: dict | None,
 ) -> EnvStats:
-    """Play episodes against a single environment and return summary stats.
+    """Play num_episodes baseline games of the model vs in-harness MCTS.
 
-    Stops early if CONSECUTIVE_FAILURE_LIMIT episodes fail in a row — the
-    remaining episodes would almost certainly fail too (model hallucinating,
-    timeouts), so there's no signal in continuing.
+    Uses the same tool-calling format as eval (core.pvp), so the baseline is
+    measured consistently with how the model is evaluated — no external server.
     """
-    seed_rng = random.Random(42)
-    scores: list[float] = []
-    consecutive_failures = 0
+    extra = eval_payload_extra or {}
+    mcts_simulations = extra.get("mcts_max_simulations")
 
-    print(f"  {env_name.value}: playing {num_episodes} episodes...", flush=True)
+    config = ChatCompletionConfig(
+        inference_model=model_name,
+        base_url=sglang_base_url,
+        temperature=ENV_EVAL_TEMPERATURE,
+    )
+    client = create_client(config)
+    chat_fn = functools.partial(chat_completion, client)
 
-    for i in range(num_episodes):
-        seed = seed_rng.randint(1, 1_000_000)
-        task_id = _sample_task_id(seed, task_id_min, task_id_max)
+    print(f"  {env_name.value}: playing {num_episodes} games vs MCTS...", flush=True)
+    result = run_mcts_baseline(
+        env_name=env_name,
+        chat_fn=chat_fn,
+        config=config,
+        num_games=num_episodes,
+        mcts_simulations=mcts_simulations,
+    )
 
-        payload: dict = {
-            "model": model_name,
-            "base_url": sglang_base_url,
-            "task_id": task_id,
-            "temperature": ENV_EVAL_TEMPERATURE,
-            "seed": seed,
-        }
-        if eval_payload_extra:
-            payload.update(eval_payload_extra)
-
-        failed = False
-        try:
-            timeout = aiohttp.ClientTimeout(total=ENV_EVAL_TASK_TIMEOUT)
-            async with session.post(
-                f"{env_server_url}/evaluate", json=payload, timeout=timeout,
-            ) as resp:
-                if resp.status == 200:
-                    data = await resp.json()
-                    result = data.get("result", data)
-                    score = float(result.get("score", 0.0))
-                else:
-                    score = 0.0
-                    failed = True
-        except Exception as e:
-            print(f"  {env_name.value} episode {i+1}: error {e}", flush=True)
-            score = 0.0
-            failed = True
-
-        scores.append(score)
-
-        if failed:
-            consecutive_failures += 1
-            if consecutive_failures >= CONSECUTIVE_FAILURE_LIMIT:
-                print(
-                    f"  {env_name.value}: {CONSECUTIVE_FAILURE_LIMIT} consecutive failures, "
-                    f"stopping early at episode {i+1}/{num_episodes}",
-                    flush=True,
-                )
-                break
-        else:
-            consecutive_failures = 0
-
+    # Per-game scores (win=1, draw=0.5, loss=0) -> the usual EnvStats summary.
+    scores = [1.0] * result.wins + [0.5] * result.draws + [0.0] * result.losses
     stats = _build_env_stats(scores)
-    print(f"  {env_name.value}: {stats.num_episodes} episodes, mean={stats.mean_score:.3f}", flush=True)
+    print(f"  {env_name.value}: {result.num_games} games, mean={stats.mean_score:.3f}", flush=True)
     return stats
 
 
@@ -218,23 +184,17 @@ async def compute_env_stats(
     try:
         await wait_for_health(sglang_local_url, "/v1/models", SGLANG_HEALTH_TIMEOUT, service_name="sglang")
 
-        print(f"SGLang ready, base_url for env servers: {sglang_base_url}", flush=True)
-        print(f"Evaluating {len(env_configs)} environments...", flush=True)
+        print(f"SGLang ready at {sglang_base_url}", flush=True)
+        print(f"Evaluating {len(env_configs)} environments vs MCTS...", flush=True)
 
-        async with aiohttp.ClientSession() as session:
-            for env_name, cfg in env_configs.items():
-                stats = await _play_episodes(
-                    session=session,
-                    env_name=env_name,
-                    env_server_url=cfg["url"],
-                    sglang_base_url=sglang_base_url,
-                    model_name=model_name,
-                    num_episodes=cfg["num_episodes"],
-                    task_id_min=cfg["task_id_min"],
-                    task_id_max=cfg["task_id_max"],
-                    eval_payload_extra=cfg.get("eval_payload_extra"),
-                )
-                all_stats[env_name] = stats
+        for env_name, cfg in env_configs.items():
+            all_stats[env_name] = _mcts_baseline_stats(
+                env_name=env_name,
+                sglang_base_url=sglang_base_url,
+                model_name=model_name,
+                num_episodes=cfg["num_episodes"],
+                eval_payload_extra=cfg.get("eval_payload_extra"),
+            )
 
     except TimeoutError:
         print("SGLang failed to start within timeout", flush=True)
